platform fix, 0 warnings

+++ src/Utilities/FileHelper.hpp
@@ -46,7 +46,7 @@ namespace Utilities
             const std::string characters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
             std::random_device rd;
             std::mt19937 generator(rd());
-            std::uniform_int_distribution<> distribution(0, characters.size() - 1);
+            std::uniform_int_distribution<> distribution(0, static_cast<int>(characters.size()) - 1);
 
             std::string randomName;
             for (size_t i = 0; i < length; ++i) {
